[green] added a new mock class. The interface files have been refactored and moved to separated files. Fixed the test.
