+++ tdd_intro/homework/07_filecopier/07_filecopier.pro
@@ -7,3 +7,7 @@ CONFIG -= qt
 
 SOURCES += \
     test.cpp
+
+HEADERS += \
+    ifilecopier.h \
+    ifilesystem.h
