+++ tdd_intro/homework/07_filecopier/ifilecopier.h
@@ -1,0 +1,6 @@
+#pragma once
+class IFileCopier
+{
+public:
+    virtual void copy(const std::string& src, const std::string& dst) = 0;
+};
