+++ tdd_intro/homework/07_filecopier/ifilesystem.h
@@ -1,0 +1,11 @@
+#pragma once
+
+using StringVector = std::vector<std::string>;
+class IFileSystem
+{
+public:
+    virtual bool isExist(const std::string& path) = 0;
+    virtual StringVector getChildList(const std::string& path) = 0;
+    virtual bool isDirectory(const std::string& path) = 0;
+    virtual void copy(const std::string& src, const std::string& dst) = 0;
+};
